+++ src/euphorie/client/browser/webhelpers.py
@@ -104,6 +104,7 @@ class WebHelpers(BrowserView):
 
     group_model = Group
     survey_session_model = SurveySession
+    dashboard_tabs = ["surveys", "assessments", "organisation"]
 
     def to_decimal(self, value):
         """Transform value in to a decimal"""
@@ -611,8 +612,8 @@ class WebHelpers(BrowserView):
         return ""
 
     def get_dashboard_tab(self):
-        tail = path.split(self.request.PATH_INFO)[-1].lstrip("@@").rstrip("/")
-        if tail in ["surveys", "assessments", "organisation"]:
+        tail = path.split(self.request.PATH_INFO.rstrip("/"))[-1].lstrip("@@")
+        if tail in self.dashboard_tabs:
             return tail
         return "dashboard"